chore: move empty response to the value objects namespace
