+++ src/Subscriptions/SubscriptionClient.php
@@ -4,6 +4,7 @@ namespace Imdhemy\GooglePlay\Subscriptions;
 
 use GuzzleHttp\Client;
 use GuzzleHttp\Exception\GuzzleException;
+use Imdhemy\GooglePlay\ValueObjects\EmptyResponse;
 use Imdhemy\GooglePlay\ValueObjects\Time;
 
 /**
